+++ src/app/health/activity/styles/ActivityHeader.module.css
@@ -1,0 +1,23 @@
+.header {
+  width: 1280px;
+  min-height: 72px;
+  background: transparent;
+}
+
+.title {
+  font-weight: 400;
+  font-size: 30px;
+  line-height: 1.2em;
+  margin-bottom: 4px;
+
+  color: #594a3e;
+}
+
+.subtitle {
+  font-weight: 400;
+  font-size: 16px;
+  line-height: 1.56em;
+  color: #4b5563;
+  margin: 0;
+  max-width: 569px;
+}
