+++ src/app/health/activity/styles/PetProfileSelector.module.css
@@ -1,0 +1,98 @@
+.petProfileSection {
+  width: 1280px;
+  min-height: 259px;
+  background: #ffffff;
+  border: 1px solid #e5e7eb;
+  border-radius: 12px;
+  box-shadow: 0px 1px 2px 0px rgba(0, 0, 0, 0.05);
+  margin-bottom: 40px;
+  overflow: hidden;
+}
+
+.tabNavigation {
+  display: flex;
+  border-top: 1px solid #e5e7eb;
+}
+
+.tab {
+  padding: 13px 24px;
+  font-weight: 400;
+  font-size: 16px;
+  line-height: 1.19em;
+  color: #4b5563;
+  cursor: pointer;
+  border-bottom: 1px solid #e5e7eb;
+  transition: all 0.3s ease;
+  min-width: 150px;
+  text-align: center;
+}
+
+.tab.active {
+  background: #f5a623;
+  color: #ffffff;
+  border-bottom: 2px solid #f5a623;
+}
+
+.tab:not(.active):hover {
+  background: #f9fafb;
+}
+
+.sectionTitle {
+  font-weight: 400;
+  font-size: 20px;
+  line-height: 1.19em;
+  color: #594a3e;
+  margin: 0;
+  padding: 30px 25px 20px;
+}
+
+.petProfiles {
+  display: flex;
+  gap: 16px;
+  padding: 0 25px 20px;
+}
+
+.petCard {
+  background: #ffffff;
+  border: 1px solid #e5e7eb;
+  border-radius: 8px;
+  padding: 18px;
+  cursor: pointer;
+  transition: all 0.3s ease;
+  min-width: 206px;
+  height: 95px;
+}
+
+.petCard.active {
+  background: #f0f5ec;
+  border: 2px solid #8bc34a;
+}
+
+.petInfo {
+  display: flex;
+  align-items: center;
+  gap: 12px;
+}
+
+.petAvatar {
+  width: 48px;
+  height: 48px;
+  border-radius: 50%;
+  object-fit: cover;
+}
+
+.petDetails h3 {
+  font-weight: 400;
+  font-size: 16px;
+  line-height: 1.5em;
+  color: #594a3e;
+  margin: 0 0 4px 0;
+}
+
+.petDetails p {
+  font-weight: 400;
+  font-size: 14px;
+  line-height: 1.43em;
+  color: #4b5563;
+  margin: 0;
+}
